C Program to Implement a Doubly Linked List
This C Program implement a doubly linked list & provide insertion, deletion & display operations. Doubly linked list is a linked data structure that consists of a set of sequentially linked records called nodes. Each node contains two fields, called links, that are references to the previous and to the next node in the sequence of nodes. The beginning and ending nodes’ previous and next links, respectively, point to some kind of terminator, typically a sentinel node or null, to facilitate traversal of the list. Here we need to preform insertion, deletion and display all the modes of given doubly linked list.
Here is source code of the C Program to implement a doubly linked list & provide insertion, deletion & display operations.